Move sorting of procedures, medications, and progression to FluxBreastCancer.
